Adding babel value tests
