+++ src/match/match.js
@@ -1,8 +1,8 @@
 import { MatchError, ParseError } from './errors.js';
-import Case from './case.js';
+import CaseBuilder from './case.js';
 
 const match = (value) => (...functionCases) => {
-  let cases = functionCases.map( aCase => new Case(aCase) )
+  let cases = functionCases.map( aCase => CaseBuilder.build(aCase) )
 
   let matchingCase = cases.find( aCase => aCase.matches(value) )
   if(!matchingCase)
